arize/requirements.txt
arize-phoenix
openinference-semantic-conventions
openai
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
python-dotenv
requests
rich
tiktoken
